refact: Person 모델 클래스 분리
